choose fig tool bar added

+++ Figures/CTriangle.cpp
@@ -38,10 +38,10 @@ double CTriangle::area(int x1, int y1, int x2, int y2, int x3, int y3)
 
 bool CTriangle:: IsInside(int x,int y)
 {
-	float A=area(vertix1.x,vertix1.y,vertix2.x,vertix2.y,vertix3.x,vertix3.y);
-	float A1=area(x,y,vertix2.x,vertix2.y,vertix3.x,vertix3.y);
-	float A2=area(vertix1.x,vertix1.y,x,y,vertix3.x,vertix3.y);
-	float A3=area(vertix1.x,vertix1.y,vertix2.x,vertix2.y,x,y);
+	double A=area(vertix1.x,vertix1.y,vertix2.x,vertix2.y,vertix3.x,vertix3.y);
+	double A1=area(x,y,vertix2.x,vertix2.y,vertix3.x,vertix3.y);
+	double A2=area(vertix1.x,vertix1.y,x,y,vertix3.x,vertix3.y);
+	double A3=area(vertix1.x,vertix1.y,vertix2.x,vertix2.y,x,y);
 			if(A==(A1+A2+A3))
 			{
 				return true;
